+++ x/leverage/types/tx_test.go
@@ -34,7 +34,7 @@ func TestTxs(t *testing.T) {
 		types.NewMsgMaxBorrow(testAddr, denom),
 		types.NewMsgRepay(testAddr, token),
 		types.NewMsgLiquidate(testAddr, testAddr, token, uDenom),
-		types.NewMsgFastLiquidate(testAddr, testAddr, token.Denom, uDenom),
+		types.NewMsgLeveragedLiquidate(testAddr, testAddr, token.Denom, uDenom),
 	}
 
 	for _, tx := range txs {
@@ -65,7 +65,7 @@ func TestRoutes(t *testing.T) {
 		types.NewMsgMaxBorrow(testAddr, denom),
 		types.NewMsgRepay(testAddr, token),
 		types.NewMsgLiquidate(testAddr, testAddr, token, uDenom),
-		types.NewMsgFastLiquidate(testAddr, testAddr, token.Denom, uDenom),
+		types.NewMsgLeveragedLiquidate(testAddr, testAddr, token.Denom, uDenom),
 	}
 
 	for _, tx := range txs {